:sparkles::fire: Deploy new details page/delete
