--- src/app/cube/old-details/components/action-panel/action-panel.component.html
@@ -1,99 +1,0 @@
-<div class="inner">
-  <div *ngIf="!isReleased && !revisedVersion" class="download-warning" tip="This learning object is pending review and may be downloaded when the review is complete."
-    tipPosition="top">
-    <i class="fas fa-exclamation"></i>Review Pending
-  </div>
-  <div *ngIf="!revisedVersion; else reviewerPanel">
-    <div class="btn-group vertical" [ngClass]="{'review-pending': !isReleased}">
-      <ng-container *ngIf="disableLibraryButtons; else libraryButtonsTemplate">
-        <div [tip]="serviceOutageMessage" class="library-buttons-disabled">
-          <ng-container *ngTemplateOutlet="libraryButtonsTemplate"></ng-container>
-        </div>
-      </ng-container>
-      <clark-editorial-action-pad *ngIf="userCanRevise" [hasRevision]="hasRevision" [learningObject]="learningObject" [revisedLearningObject]="revisedLearningObject"></clark-editorial-action-pad>
-    </div>
-    <ng-container *ngIf="!auth.user">
-      <div tabindex="0" class="action-message" id="login-msg">Please
-        <a [routerLink]="['/auth/login']">log in</a> to add this Learning Object to your library.
-      </div>
-    </ng-container>
-    <ng-container *ngIf="auth.user && !auth.user?.emailVerified">
-      <div tabindex="0" class="action-message">
-        Please <a (activate)="sendEmailVerification()">verify your email</a> to download this Learning Object.
-      </div>
-    </ng-container>
-    <div class="details-metrics">
-      <div tabindex="0">
-        <i class="far fa-save"></i>
-        <span #savesRef>{{ learningObject?.metrics.saves }}</span> save{{ (learningObject?.metrics.saves) !== 1 ? 's' : '' }}
-      </div>
-      <div tabindex="0">
-        <i class="far fa-download"></i>
-        <span>{{ learningObject?.metrics.downloads }}</span> download{{ (learningObject?.metrics.downloads) !== 1 ? 's' : '' }}
-      </div>
-    </div>
-    <div class="cc-attribution">
-      <div class="cc-attribution__text-content">
-        <div tabindex="0" class="cc-attribution__title"></div>
-        <div tabindex="0" class="cc-attribution__attribution" #objectAttributionElement id="objectAttribution"> 
-          <a attr.aria-label="Clickable link to {{ learningObject.name | titlecase }}" href='https://clark.center/details/{{learningObject.author.username}}/{{learningObject.cuid}}'>"{{ learningObject.name }}"</a>
-          by <a attr.aria-label="Clickable link to {{ learningObject.author.name | titlecase }} profile" href='https://clark.center/users/{{learningObject.author.username}}'>{{ learningObject.author.name | titlecase }}</a>
-          is licensed under <a attr.aria-label="Clickable link to Creative Commons License" href='https://creativecommons.org/licenses/by-nc-sa/4.0/'>CC BY-NC-SA 4.0</a>
-        </div>
-      </div>
-      <div class="cc-attribution__button">
-        <button aria-label="Clickable Copy CC Attribution button" class="button icon-only green" (activate)="copyAttribution()" id="copyAttribution">
-          <i class="fas fa-clipboard"></i>
-        </button>
-      </div>
-    </div>
-    <div class="details-share">
-      <div class="share-title"></div>
-      <div class="share-icons">
-        <button (activate)="shareButton($event, 'facebook')" class="button icon-only facebook" id="facebook" aria-label="Clickable Share on Facebook button">
-          <i class="fab fa-facebook"></i>
-        </button>
-        <button (activate)="shareButton($event, 'twitter')" class="button icon-only twitter" id="twitter" aria-label="Clickable Share on Twitter button">
-          <i class="fab fa-twitter"></i>
-        </button>
-        <button (activate)="shareButton($event, 'linkedin')" class="button icon-only linkedin" id="linkedin" aria-label="Clickable Share on LinkedIn button">
-          <i class="fab fa-linkedin"></i>
-        </button>
-        <button (activate)="shareButton($event, 'email')" class="button icon-only email" id="email" aria-label="Clickable Share via email button">
-          <i class="fas fa-envelope"></i>
-        </button>
-        <button (activate)="shareButton($event, 'copyLink')" class="button icon-only copylink" id="copylink" aria-label="Clickable Copy link button">
-          <i class="fas fa-clipboard"></i>
-        </button>
-        <textarea tabindex="-1" aria-label="link copy area" #objectLinkElement readonly="true" value="{{url}}" style="position: fixed; opacity: 0; width: 0; height:0;"></textarea>
-      </div>
-    </div>
-  </div>
-  <ng-template #reviewerPanel>
-    <clark-reviewer-panel [revisedDate]="revisedDate" [releasedDate]="releasedDate" (download)="downloadRevised($event)"></clark-reviewer-panel>
-  </ng-template>
-</div>
-
-<clark-popup *ngIf="showDownloadModal" (closed)="toggleDownloadModal(false)">
-  <div style="max-width: 600px;" #popupInner>
-    <clark-download-notice-popup (close)="toggleDownloadModal(false)"></clark-download-notice-popup>
-  </div>
-</clark-popup>
-
-<ng-template #libraryButtonsTemplate>
-  <button [tip]="tips.DOWNLOAD_NOW" tipPosition="top" [tipDisabled]="hasDownloadAccess || !loggedin" [ngClass]="{'disabled': !auth.user || !hasDownloadAccess || downloading || disableLibraryButtons}"
-    class="button good" id="download-button" (activate)="addToCart(true)" attr.aria-label="{{hasDownloadAccess ? 'Clickable Download Now button' : 'Not Available for Download'}}">
-    <span *ngIf="hasDownloadAccess">{{!downloading ? 'Download Now' : 'Downloading'}}
-      <span *ngIf="downloading">
-        <i class="fal fa-spinner-third fa-spin"></i>
-      </span>
-    </span>
-    <span tabindex="0" *ngIf="!hasDownloadAccess">Not available for download</span>
-  </button>
-  <button *ngIf="learningObject?.status === 'released' && !userIsAuthor" [ngClass]="{'disabled': !auth.user || disableLibraryButtons, 'saved': saved}" class="button neutral" id="save-to-library" (activate)="addToCart()" attr.aria-label="{{ saved ? 'This Learning Object is Saved to your library!' : 'Clickable Save to Library button' }}">
-    {{ saved ? 'Saved to your library!' : 'Save to Library' }}
-    <span *ngIf="addingToLibrary">
-      <i class="fal fa-spinner-third fa-spin"></i>
-    </span>
-  </button>
-</ng-template>
