--- src/app/cube/old-details/components/action-panel/action-panel.component.scss
@@ -1,229 +1,0 @@
-@import '~_vars.scss';
-
-.inner {
-  background: white;
-  border-radius: 2px;
-  box-shadow: 0 0 12px 1px rgba(0, 0, 0, 0.03);
-  padding: 25px;
-  margin-bottom: 20px;
-
-  .download-warning {
-    color: white;
-    background: $orange-gradient;
-    padding: 8px 12px;
-    border-radius: 2px 2px 0 0;
-    font-size: $large;
-    text-align: center;
-    margin-bottom: 10px;
-    text-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-
-    .svg-inline--fa {
-      margin-right: 10px;
-    }
-  }
-
-  .library-buttons-disabled {
-    width: 100%;
-    margin-bottom: 20px;
-
-    & > *.button {
-      width: 100%;
-    }
-  }
-
-  .btn-group.review-pending {
-    margin-top: 40px;
-
-    & > * {
-      display: block;
-      margin-bottom: 15px;
-    }
-  }
-
-  .button {
-    overflow: hidden;
-    text-align: center;
-    position: relative;
-
-    &.saved {
-      cursor: default !important;
-      pointer-events: none !important;
-      color: $success-green;
-
-      &:after {
-        border-color: transparent ;
-      }
-    }
-  }
-
-  .action-message {
-    margin-top: 20px;
-    font-size: $normal;
-    color: $dark-grey;
-
-    a {
-      color: $blue-accent;
-      cursor: pointer;
-    }
-  }
-}
-
-.details-metrics {
-  font-size: $normal;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  color: $dark-grey;
-  margin-top: 30px;
-
-  & > * {
-    span {
-      position: relative;
-
-      &.animate {
-        animation: newSaveOrDownload 0.8s ease;
-      }
-    }
-
-    &:first-child {
-      margin-right: 20px;
-    }
-  }
-
-  .svg-inline--fa {
-    margin-right: 5px;
-  }
-}
-
-.cc-attribution {
-  margin-top: 50px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  position: relative;
-  font-size: $small;
-
-  &:before {
-    content: 'Attribute this Object';
-    color: $dark-grey;
-    text-align: center;
-    padding: 0;
-    background: white;
-    position: absolute;
-    z-index: 2;
-    left: 20%;
-    right: 20%;
-    top: -30px
-  }
-
-  &:after {
-    content: '';
-    position: absolute;
-    height: 1px;
-    background: $light-grey;
-    width: 100%;
-    left: 0;
-    top: -20px;
-  }
-
-  .cc-attribution__title {
-    position: relative;
-  }
-
-  .attribution__attribution {
-    padding-top: 20px;
-    width: 80%;
-    display: inline-block;
-  }
-
-  .cc-attribution__button {
-    min-width: 50px;
-    display: flex;
-    justify-content: flex-end;
-  }
-}
-
-.details-share {
-  margin-top: 30px;
-
-  .share-title {
-    position: relative;
-
-    &:before {
-      content: 'Share';
-      color: $dark-grey;
-      text-align: center;
-      padding: 0 10px;
-      background: white;
-      position: absolute;
-      z-index: 2;
-      left: 30%;
-      right: 30%;
-      bottom: -7px;
-    }
-
-    &:after {
-      content: '';
-      position: absolute;
-      height: 1px;
-      background: $light-grey;
-      width: 100%;
-      left: 0;
-      bottom: 0;
-    }
-  }
-
-  .share-icons {
-    padding-top: 20px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-
-    a {
-      color: white !important;
-      text-decoration: none !important;
-    }
-
-    & > button {
-      background: $blue-accent;
-      color: white !important;
-      transform: scale(1);
-      cursor: pointer;
-
-      &:active:hover {
-        transform: scale(0.95);
-      }
-
-      &.facebook {
-        background: #3b5998;
-      }
-
-      &.twitter {
-        background: #1da1f2;
-      }
-
-      &.linkedin {
-        background: #0077b5;
-      }
-
-      &.email {
-        background: $error-red;
-      }
-
-      &.copylink {
-        background: $success-green;
-      }
-    }
-  }
-}
-
-.email-unverified {
-  background: $error-red;
-  color: white;
-  padding: 3px;
-  border-radius: 2px;
-}
