--- src/app/cube/old-details/components/action-panel/editorial-action-pad/editorial-action-pad.component.scss
@@ -1,34 +1,0 @@
-@import '~_vars.scss';
-
-button.neutral {
-  width: 100%;
-}
-
-p {
-  font-size: $normal;
-}
-.btn-group {
-  margin-top: 50px;
-
-  button {
-    width: auto;
-  }
-}
-
-.popup-content {
-  max-width: 600px;
-  padding: 16px;
-}
-
-.popup-header {
-  width: 100%;
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  margin-bottom: 25px;
-
-  .fa-exclamation-triangle {
-    font-size: 60px;
-    color: $medium-grey;
-  }
-}
