--- src/app/cube/old-details/components/children-detail-view/children-detail-view.component.html
@@ -1,6 +1,0 @@
-<section class="children-detail-view">
-  <div  class="header">
-    <h1 tabindex="0">This {{ length | titlecase }} includes...</h1>
-  </div>
-  <learning-object-component *ngFor="let child of children" [learningObject]="child"></learning-object-component>
-</section>
