--- src/app/cube/old-details/components/children-detail-view/children-detail-view.component.scss
@@ -1,16 +1,0 @@
-@import '~_vars.scss';
-
-.children-detail-view {
-  
-  h1 {
-    color: $darker-grey;
-    font-size: $larger;
-    padding-left: 10px;
-    font-weight: normal;
-  }
-
-  learning-object-component:not(:last-child) {
-    display: block;
-    margin-bottom: 10px;
-  }
-}
