--- src/app/cube/old-details/components/error-status/error-status.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ErrorStatusComponent } from './error-status.component';
-
-describe('ErrorStatusComponent', () => {
-  let component: ErrorStatusComponent;
-  let fixture: ComponentFixture<ErrorStatusComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ErrorStatusComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ErrorStatusComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
