--- src/app/cube/old-details/components/file-details/file-details.component.html
@@ -1,40 +1,0 @@
-<section class="file-details-wrapper" *ngIf="!(!materials.files.length && !materials.urls.length && !materials.notes.length)">
-  <div tabindex="0" class="section title spaced">
-    {{ copy.TITLEPREFIX }} {{ length | titlecase}}
-  </div>
-  <div class="file-details-wrapper section">
-    <section class="files" *ngIf="materials.files.length">
-      <div class="grid-area" [ngClass]="{header: materials.files.length}">
-        <div tabindex="0" class="section title">
-          {{ copy.FILESANDFOLDERS }}
-        </div>
-        <ng-container>
-          <div tabindex="0" class="description">
-            {{ materials.files.length }} {{ copy.FILES }}
-          </div>
-        </ng-container>
-      </div>
-
-
-      <div class="file-listing">
-        <clark-file-browser [files$]="files$" [folderMeta$]="folderMeta$"></clark-file-browser>
-      </div>
-
-    </section>
-
-    <section class="links" *ngIf="materials.urls.length > 0">
-      <div tabindex="0" class="section title">{{ copy.LINKS }}</div>
-      <div class="link-listing" *ngFor="let link of materials.urls">
-        <span tabindex="0" class="link-listing-title">{{ link.title }}</span>
-        <span class="link-listing-url">
-          <a href="{{ link.url }}" target="_blank" attr.aria-label="Clickable link to {{ link.title }}">{{ copy.OPENLINK }}</a>
-        </span>
-      </div>
-    </section>
-
-    <section class="notes" *ngIf="materials.notes.length > 0">
-      <div tabindex="0" class="section title">{{ copy.NOTES }}</div>
-      <p tabindex="0">{{ materials.notes }}</p>
-    </section>
-  </div>
-</section>
