--- src/app/cube/old-details/components/file-details/file-details.component.scss
@@ -1,118 +1,0 @@
-@import '~_vars.scss';
-
-.file-details-wrapper {
-  background: white;
-  padding: 20px;
-  border-radius: 2px;
-  color: $dark-grey;
-  margin-bottom: 40px;
-}
-
-.header {
-  margin-bottom: 20px;
-}
-
-.files,
-.links,
-.notes {
-  margin-bottom: 40px;
-
-  p {
-    padding-left: 10px;
-    padding-top: 8px;
-  }
-}
-
-.section.title {
-  color: $darker-grey;
-  font-size: $larger;
-
-  .spaced {
-    margin-bottom: 15px;
-  }
-}
-
-::ng-deep .grid-area {
-  display: grid;
-  grid-template-columns: 2fr repeat(2, 1fr);
-  grid-template-areas: 'title file-count description';
-  align-items: center;
-
-  .title {
-    display: flex;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    grid-area: title;
-  }
-
-  .file-count {
-    grid-area: file-count;
-  }
-
-  .description {
-    display: grid;
-    justify-content: flex-end;
-    grid-area: description;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-}
-
-.file-listing {
-  display: grid;
-  grid-row-gap: 10px;
-  margin: 20px 0;
-  grid-auto-columns: 100%;
-}
-
-.link-listing {
-  word-wrap: break-word;
-}
-
-.link-button {
-  color: $light-blue;
-  cursor: pointer;
-  transition: opacity 0.15s ease;
-  -webkit-transition: opacity 0.15s ease;
-}
-
-.link-button:active {
-  opacity: 0.3;
-  transition: opacity 0.15s ease;
-  -webkit-transition: opacity 0.15s ease;
-}
-
-.links {
-  $section-width: 45%;
-
-  .link-listing {
-    display: flex;
-    flex-flow: row nowrap;
-    padding: 8px 0;
-
-    .link-listing-title,
-    .link-listing-url {
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-
-    .link-listing-title {
-      width: $section-width;
-      flex-grow: 1;
-      margin-left: 1%;
-      margin-right: 8%;
-    }
-
-    .link-listing-url {
-      margin-right: 1%;
-    }
-  }
-}
-
-
-// mobile styles
-@media screen and (max-width: "600px") {
-  .section.title {
-    font-size: $large;
-  }
-}
