--- src/app/cube/old-details/components/file-details/file-details.copy.ts
@@ -1,10 +1,0 @@
-export const COPY = {
-    TITLEPREFIX: 'Materials for this',
-    FILESANDFOLDERS: 'Files and Folders',
-    FILES: 'Files',
-    FILESIZE: 'Filesize',
-    LINKS: 'Links',
-    OPENLINK: 'Open Link',
-    NOTES: 'Notes',
-
-};
