--- src/app/cube/old-details/components/included-in/included-in.component.scss
@@ -1,18 +1,0 @@
-@import '~_vars.scss';
-
-.included-in {
-  border-radius: 2px;
-  box-shadow: 0 0 12px 1px rgba(0, 0, 0, 0.03);
-
-  .header {
-    border-bottom: $light-grey;
-    padding: 8px;
-    font-size: $large;
-    background-color: white;
-
-    .lo-length {
-      text-transform: capitalize;
-    }
-  }
-
-}
