--- src/app/cube/old-details/components/included-in/parent-listing.component.scss
@@ -1,42 +1,0 @@
-@import '~_vars.scss';
-
-.parent-listing {
-  padding: 10px;
-  background-color: white;
-
-  .type {
-    font-size: $small;
-    color: white;
-    padding: 2px 4px;
-    text-transform: capitalize;
-  } // FIXME: These styles aren't DRY
-  &.nanomodule {
-    .type {
-      background: $nanomodule-color;
-    }
-  }
-
-  &.micromodule {
-    .type {
-      background: $micromodule-color;
-    }
-  }
-
-  &.module {
-    .type {
-      background: $module-color;
-    }
-  }
-
-  &.unit {
-    .type {
-      background: $unit-color;
-    }
-  }
-
-  &.course {
-    .type {
-      background: $course-color;
-    }
-  }
-}
