--- src/app/cube/old-details/components/learning-object-ratings/learning-object-ratings.component.html
@@ -1,82 +1,0 @@
-<div class="learning-object-ratings">
-  <div tabindex="0" class="title">
-    Reviews
-  </div>
-  <div class="overall-rating">
-    <div *ngIf="ratings?.length > 0">Overall Rating</div>
-    <clark-rating-stars  *ngIf="ratings?.length > 0" [rating]="averageRating" [color]="'gold'" [size]="22" [reviewsCount]="ratings?.length"></clark-rating-stars>
-  </div>
-  <div *ngIf="ratings?.length > 0; else noRatings" class="ratings-list">
-    <div *ngFor="let rating of ratings; let i = index;" class="rating-list-element">
-      <a [routerLink]="['/users', rating.user.username]">
-        <div class="rating-user">
-          <img [src]="userService.getGravatarImage(rating.user.email, 45)" class="user-image" alt="Picture of {{rating.user.name}}"
-          />
-          <div class="user-name">
-            <span>{{ rating.user.name }}</span>
-            <!-- TODO implement check for verified user here -->
-            <div *ngIf="false" class="verified"><i class="far fa-check-circle"></i>verified user</div>
-          </div>
-        </div>
-      </a>
-      <clark-rating-stars [rating]="rating.value" [color]="'gold'" [size]="18"></clark-rating-stars>
-      <div tabindex="0" class="rating-date">{{ rating.date | date:'shortDate' }}</div>
-      <div tabindex="0" class="rating-text">
-        {{ rating.comment }}
-      </div>
-      <div *ngIf="auth.user?.emailVerified" class="options">
-        <ng-container *ngIf="isRatingAuthor(i); else notOwnerOptions">
-          <span tabindex="0" aria-label="Clickable Edit Rating link" (activate)="submitEditRating(i)"><i class="far fa-pencil-alt"></i>Edit rating</span>
-          <span tabindex="0" aria-label="Clickable Delete Rating link" (activate)="submitDeleteRating(i)"><i class="far fa-trash-alt"></i>Delete rating</span>
-        </ng-container>
-        <ng-template #notOwnerOptions>
-          <span tabindex="0" aria-label="Clickable Respond to Rating link" *ngIf="isLearningObjectOwner() && rating.response.length === 0" (activate)="openResponse(i);"><i class="far fa-comment-alt"></i>Respond</span>
-          <span tabindex="0" aria-label="Clickable Report Rating link" *ngIf="loggedIn" class="report" (activate)="reportIndex = i; showReport = true;"><i class="fas fa-flag"></i>Report rating</span>
-        </ng-template>
-      </div>
-      <div *ngIf="rating.response.length > 0 && !isEditingResponse(i)" class="response-container">
-          <a [routerLink]="['/users', rating.response[0].user.username]">
-            <div class="rating-user">
-              <img [src]="userService.getGravatarImage(rating.response[0].user.email, 45)" class="user-image" alt="Picture of {{rating.response[0].user.name}}"
-              />
-              <div class="user-name">
-                <span tabindex="0">{{ rating.response[0].user.name }}</span>
-              </div>
-            </div>
-          </a>
-          <div class="rating-date" tabindex="0">{{ rating.response[0].date | date:'shortDate' }}</div>
-          <div class="rating-text" tabindex="0">
-            {{ rating.response[0].comment }}
-          </div>
-          <div class="options">
-              <ng-container *ngIf="isResponseAuthor(i); else notOwnerOptions">
-                <button tabindex="0" (activate)="openEditResponse($event, i)"><i class="far fa-pencil-alt"></i>Edit response</button>
-                <button (activate)="submitDeleteResponse($event, i)"><i class="far fa-trash-alt"></i>Delete response</button>
-              </ng-container>
-              <div *ngIf="rating.user.username === auth.username" class="email">
-                <a class="mailto" href="mailto: {{ rating.response[0].user.email }}"> Email: {{ rating.response[0].user.email }}</a>
-              </div>
-          </div>
-      </div>
-      <div *ngIf="isWritingResponse(i)">
-          <clark-new-rating-response [response]="{comment: '', index: i}" (submit)="submitResponse($event)" (cancel)="cancelResponse($event)" ></clark-new-rating-response>
-      </div>
-      <div *ngIf="isEditingResponse(i)">
-          <clark-new-rating-response [response]="{comment: rating.response[0].comment, index: i}"  (submit)="submitEditResponse($event)" (cancel)="cancelEditResponse($event)" ></clark-new-rating-response>
-      </div>
-    </div>
-  </div>
-
-  <ng-template #noRatings>
-    <div tabindex="0" class="no-ratings">
-      This learning object doesn't have any ratings yet!
-    </div>
-  </ng-template>
-</div>
-<!-- This wrapper is styled from parent component -->
-<div *ngIf="showReport" class="report-rating-wrapper" [ngClass]="{'active': showReport}" (activate)="showReport = false;">
-  <div class="report-inner" (activate)="$event.stopPropagation();">
-    <div (activate)="showReport = false;" class="closePopup"><i class="fal fa-times"></i></div>
-    <clark-report-rating (cancel)="showRating = false;" (submit)="triggerReportRating($event)"></clark-report-rating>
-  </div>
-</div>
