--- src/app/cube/old-details/components/learning-object-ratings/learning-object-ratings.component.scss
@@ -1,163 +1,0 @@
-@import '~_vars.scss';
-
-.title {
-  font-size: $large;
-  color: $darker-grey;
-}
-
-.mailto {
-  color: blue;
-  text-decoration: underline;
-}
-
-.response-container {
-  margin-left: 30px;
-  margin-top: 20px;
-}
-
-.overall-rating {
-  margin: 10px auto 40px;
-  font-size: $normal;
-}
-
-.ratings-list {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-
-  & > *:not(:last-child) {
-    border-bottom: 1px solid $light-grey;
-    padding-bottom: 25px;
-    margin-bottom: 15px;
-  }
-
-  .rating-list-element {
-    padding: 10px 0;
-
-    a {
-      color: $dark-grey;
-      text-decoration: none;
-    }
-
-    .rating-user {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-      margin-bottom: 10px;
-
-      img {
-        border-radius: 50%;
-
-        &:after {
-          content: '';
-          display: block;
-          padding-bottom: 100%;
-        }
-      }
-
-      .user-name {
-        font-size: $large;
-        text-transform: capitalize;
-        margin-left: 10px;
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-      }
-    }
-
-    .rating-date {
-      color: $dark-grey;
-      opacity: 0.8;
-      margin-top: 5px;
-      margin-bottom: 10px;
-    }
-
-    .rating-text {
-      font-size: $normal;
-    }
-  }
-
-  .options {
-    margin-top: 5px;
-    font-size: $smaller;
-
-    & > * {
-      cursor: pointer;
-      color: $dark-grey;
-
-      &:not(:last-child) {
-        margin-right: 15px;
-      }
-
-      &:hover {
-        text-decoration: underline;
-      }
-
-      &.report {
-        color: $error-red;
-      }
-
-      .svg-inline--fa {
-        margin-right: 5px;
-      }
-    }
-  }
-}
-
-.no-ratings {
-  text-align: left;
-  color: $dark-grey;
-  font-size: $large;
-}
-
-.report-rating-wrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: $dark-overlay;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  opacity: 0;
-  visibility: hidden;
-  transition: all 0.2s ease;
-  z-index: 4;
-
-  .report-inner {
-    transform: scale(0.9);
-    background: white;
-    padding: 20px;
-    transition-delay: 0s;
-    transition: all 0.2s ease;
-    opacity: 0;
-    max-width: 600px;
-    min-height: 300px;
-    width: 100%;
-    border-radius: 2px;
-    position: relative;
-
-    .closePopup {
-      position: absolute;
-      font-size: $large;
-      right: 20px;
-      top: 23px;
-      color: $blue-accent;
-      cursor: pointer;
-      z-index: 2;
-    }
-  }
-
-  &.active {
-    opacity: 1;
-    visibility: visible;
-
-    .report-inner {
-      transition-delay: 0.1s;
-      opacity: 1;
-      transform: scale(1);
-    }
-  }
-}
