--- src/app/cube/old-details/components/new-rating-response/new-rating-response.component.scss
@@ -1,13 +1,0 @@
-@import '~_vars.scss';
-
-.text-area {
-  resize: none;
-  appearance: none;
-  outline: 0;
-  border: 1px solid $medium-grey;
-  padding: 10px;
-  height: 150px;
-  width: 100%;
-  margin-top: 20px;
-  margin-bottom: 10px;
-}
