--- src/app/cube/old-details/components/new-rating/new-rating.component.html
@@ -1,24 +1,0 @@
-<div class="new-rating" [ngClass]="{'rated': rating.value && activePanel === 1}">
-  <div class="stars-wrapper" [ngClass]="{'active': activeHover > -1}" (mouseleave)="reset()">
-    <div class="empty-stars stars">
-      <div class="star" *ngFor="let item of iterableCount; let i = index;" [tip]="tips[i]" tipPosition="bottom">
-        <div (mouseover)="setHover(i)" (activate)="rate(i + 1)">
-          <i class="fal fa-star"></i>
-        </div>
-      </div>
-    </div>
-    <div class="full-stars stars">
-      <div *ngFor="let item of iterableCount; let i = index;" class="star" [ngClass]="{'show': starShouldShow(i)}">
-        <i class="fas fa-star"></i>
-      </div>
-    </div>
-  </div>
-  <div class="comment-wrapper">
-    <textarea aria-label="comment" placeholder="Why did you leave this rating?" [(ngModel)]="rating.comment" userscaleable></textarea>
-    <div class="btn-group center">
-      <button class="button good" [ngClass]="{'disabled': !rating.comment || rating.comment === ''}" (activate)="submitRating()">Submit
-        <i class="far fa-arrow-right"></i></button>
-      <button class="button neutral" (activate)="cancel()">Cancel <i class="far fa-ban"></i></button>
-    </div>
-  </div>
-</div>
