--- src/app/cube/old-details/components/new-rating/new-rating.component.scss
@@ -1,90 +1,0 @@
-@import '~_vars.scss';
-
-.new-rating {
-  &.rated {
-    .comment-wrapper {
-      visibility: visible;
-      opacity: 1;
-      transform: scale(1);
-      max-height: 220px;
-      margin: 20px auto;
-    }
-  }
-
-  .stars-wrapper {
-    position: relative;
-    max-width: 200px;
-    margin: auto;
-
-    .stars {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      font-size: 30px;
-      width: 200px;
-      color: $blue-accent;
-      z-index: 1;
-      position: relative;
-
-      &.full-stars {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        z-index: 0;
-      }
-
-      .star {
-        cursor: pointer;
-        transition: all 0.2s ease;
-
-        &.show {
-          visibility: visible;
-          transform: scale(1);
-        }
-      }
-    }
-
-    .full-stars .star {
-      visibility: hidden;
-      transform: scale(0);
-      transition: all 0.2s ease;
-    }
-
-    &.active .stars {
-      color: $light-blue;
-    }
-  }
-
-  .comment-wrapper {
-    transform: scale(0.7);
-    opacity: 0;
-    visibility: hidden;
-    transition: all 0.6s ease;
-    max-height: 0;
-    overflow: hidden;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    textarea {
-      resize: none;
-      appearance: none;
-      outline: 0;
-      border: 1px solid $medium-grey;
-      padding: 10px;
-      height: 150px;
-      width: 350px;
-
-      &:focus {
-        border-color: $blue-accent;
-      }
-    }
-
-    .btn-group {
-      margin-top: 20px;
-      margin-bottom: 5px;
-    }
-  }
-}
