--- src/app/cube/old-details/components/outcomes-detail-view/outcomes-detail-view.component.scss
@@ -1,99 +1,0 @@
-@import '~_vars.scss';
-
-.outcomes-detail-view {
-  margin-bottom: 40px;
-
-  h1 {
-    color: $darker-grey;
-    font-size: $larger;
-    padding-left: 10px;
-    font-weight: normal;
-  }
-
-  .no-content-alert {
-    margin-left: 10px;
-    color: $dark-grey;
-  }
-
-  .outcomes {
-    border: 1px solid $light-grey;
-    border-radius: 2px;
-    background: white;
-    color: $dark-grey;
-
-    .header {
-      display: flex;
-      flex-flow: row nowrap;
-      align-items: center;
-      width: 100%;
-      border: 0;
-      background: white;
-
-      &:focus {
-        outline: 0;
-      }
-
-      .title {
-        flex-grow: 2;
-        color: $darker-grey;
-        font-size: $small;
-        padding: 10px;
-        padding-left: 20px;
-        text-align: start;
-      }
-
-      .mappings-count {
-        font-size: $small;
-        color: $dark-grey;
-        padding-right: 10px;
-      }
-
-      &:not(.clickable) {
-        cursor: default;
-      }
-
-      &.clickable {
-        &:hover {
-          outline: 1px solid $light-blue;
-        }
-      }
-    }
-
-    ul {
-      border-top: 1px solid $light-blue;
-      display: flex;
-      flex-direction: column;
-      flex-wrap: nowrap;
-      justify-content: space-between;
-      padding-right: 40px;
-    }
-
-    li {
-      margin-bottom: 25px;
-      list-style-type: none; // Grid layout
-      display: grid;
-      grid-template-columns: auto 175px; //FIXME: Remove hardcoded col 2 value
-      grid-row-gap: 10px;
-
-      .standard-author,
-      .standard-name {
-        font-size: $normal;
-        color: $dark-grey;
-      }
-
-      .standard-name {
-        text-align: right;
-      }
-
-      .standard-outcome {
-        font-size: $small;
-      }
-
-      .standard-author,
-      .standard-name,
-      .standard-outcome {
-        padding: 8px;
-      }
-    }
-  }
-}
