--- src/app/cube/old-details/components/report-rating/report-rating.component.html
@@ -1,32 +1,0 @@
-<div class="report-rating">
-  <div class="title">
-    <div *ngIf="page === 2" (activate)="page = 1" class="icon"><i class="far fa-chevron-left"></i></div>
-    Report This {{ subject | titlecase }}
-  </div>
-  <ng-container *ngIf="page === 1; else commentTemplate">
-    <div class="type-list">
-      <ul>
-        <li (activate)="selectType(i)" class="type" *ngFor="let t of types; let i = index;">
-          {{ t }}
-          <i class="far fa-chevron-right"></i>
-        </li>
-      </ul>
-    </div>
-  </ng-container>
-  <ng-template #commentTemplate>
-    <div class="comment">
-      <div class="comment-description">
-        <ng-container *ngIf="type !== undefined && type <= 1">
-          You're reporting this {{ subject }} because <span>{{ types[type] | lowercase }}</span>.
-        </ng-container>
-        Can you tell us more about why
-        <ng-container *ngIf="type !== undefined && type === 2"> you're reporting this {{ subject }}</ng-container>?
-      </div>
-      <textarea aria-label="report comment" [(ngModel)]="comment" placeholder="More information..."></textarea>
-      <div class="btn-group center">
-        <button class="button bad" (activate)="triggerReport()">Send Report<i class="fas fa-flag"></i></button>
-        <button class="button neutral">Cancel<i class="far fa-ban"></i></button>
-      </div>
-    </div>
-  </ng-template>
-</div>
