--- src/app/cube/old-details/components/report-rating/report-rating.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ReportRatingComponent } from './report-rating.component';
-
-describe('ReportRatingComponent', () => {
-  let component: ReportRatingComponent;
-  let fixture: ComponentFixture<ReportRatingComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ReportRatingComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ReportRatingComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
