--- src/app/cube/old-details/components/reviewer-panel/reviewer-panel.component.scss
@@ -1,22 +1,0 @@
-@import '~_vars.scss';
-
-.button {
-  overflow: hidden;
-  text-align: center;
-  font-size: $normal;
-  margin-bottom: 10px;
-  border: none;
-  position: relative;
-  padding-top: 10px;
-  padding-bottom: 10px;
-
-  &.good {
-    background: $orange !important;
-  }
-}
-.last-revised,
-.last-released {
-  margin-bottom: 10px;
-  color: $dark-grey;
-  font-size: $normal;
-}
