--- src/app/cube/old-details/components/splash/details-splash.component.scss
@@ -1,177 +1,0 @@
-@import '~_vars.scss';
-
-$sidebar-width: 320px;
-$content-width: calc(#{$max-width} - #{$sidebar-width});
-
-.splash {
-  background: $blue-gradient;
-  padding: 50px 0;
-  position: relative;
-
-  &:after {
-    content: '';
-    background: white;
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    opacity: 0.03;
-    mask: $cube-pattern repeat;
-  }
-
-  .inner {
-    max-width: $max-width;
-    margin: auto;
-    padding: 20px;
-    box-sizing: border-box;
-    z-index: 1;
-
-    .header {
-      max-width: $content-width;
-      padding-right: 20px;
-      padding-bottom: 60px;
-
-      .collection {
-        position: absolute;
-        margin-top: 30px;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-
-        .collection-image {
-          position: relative;
-          width: 60px;
-          background-size: contain;
-          background-position: center;
-          background-repeat: no-repeat;
-
-          &:after {
-            content: '';
-            display: block;
-            padding-bottom: 100%;
-          }
-        }
-
-        .collection-name {
-          color: white;
-          font-size: $large;
-          margin-left: 10px;
-          opacity: 0.6;
-        }
-      }
-      .collection-link {
-        cursor: pointer;
-      }
-      .collection-link:hover {
-        opacity: .6;
-      }
-
-      .author {
-        .link {
-          color: white;
-          opacity: 1;
-          font-weight: 400;
-        }
-
-        .link:hover, .link:focus {
-          cursor: pointer;
-          text-decoration: underline;
-        }
-      }
-    }
-
-    .name {
-      font-size: $largest;
-      color: white;
-      text-shadow: 0 0 rgba(0, 0, 0, 0.2);
-    }
-
-    .author,
-    .downloads {
-      color: white;
-      font-size: $large;
-      font-weight: 300;
-    }
-
-    .updated {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-      
-      a {
-        font-size: $normal;
-        color: white;
-        cursor: pointer;
-        margin-left: 10px;
-        text-decoration: underline;
-        font-weight: 400;
-      }
-    }
-
-    .rating {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-
-      a {
-        font-size: $normal;
-        color: white;
-        cursor: pointer;
-        margin-left: 10px;
-        text-decoration: underline;
-      }
-
-      .emailNotVerified {
-        color: white;
-        opacity: 0.6;
-      }
-    }
-
-    .type {
-      margin-top: 5px;
-      padding: 3px 8px;
-      border-radius: 3px;
-      display: inline-block;
-      position: relative;
-      color: white;
-      text-transform: capitalize;
-      font-size: $normal;
-      font-weight: 300;
-
-      &.nanomodule {
-        background: $nanomodule-color;
-      }
-
-      &.micromodule {
-        background: $micromodule-color;
-      }
-
-      &.module {
-        background: $module-color;
-      }
-
-      &.unit {
-        background: $unit-color;
-      }
-
-      &.course {
-        background: $course-color;
-      }
-
-      &:before {
-        content: '';
-        border-radius: 50%;
-        width: 10px;
-        height: 10px;
-        left: 0;
-        top: 0;
-        bottom: 0;
-        margin: auto;
-        position: absolute;
-      }
-    }
-  }
-}
