--- src/app/cube/old-details/details.component.scss
@@ -1,308 +1,0 @@
-@import '~_vars.scss';
-
-$sidebar-width: 360px;
-$content-width: calc(#{$max-width} - #{$sidebar-width});
-
-.details-wrapper {
-  width: 100%;
-  padding-bottom: 20px;
-
-  .flex {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    flex-shrink: 0;
-    max-width: $max-width;
-    margin: auto;
-
-    .content {
-      max-width: $content-width;
-      padding: 20px;
-      flex-shrink: 1;
-      flex-grow: 1;
-      height: auto;
-
-      cube-children-detail-view {
-        display: block;
-        margin-bottom: 50px;
-      }
-
-      .description,
-      .outcomes {
-        border: 1px solid $light-grey;
-        padding: 20px;
-        border-radius: 2px;
-        color: $dark-grey;
-        margin-bottom: 40px;
-
-        .title {
-          color: $darker-grey;
-          font-size: $larger;
-          margin-bottom: 20px;
-          padding-left: 10px;
-        }
-      }
-
-      .description {
-        background: white;
-        box-shadow: 1px 1px 10px 0 rgba(0, 0, 0, 0.05);
-      }
-
-      .academic-levels {
-        display: flex;
-        flex-wrap: wrap;
-        flex-direction: row;
-        justify-content: flex-start;
-        align-items: center;
-        margin-bottom: 5px;
-
-        .title {
-          width: 100%;
-          color: $darker-grey;
-          font-size: $larger;
-          padding-left: 10px;
-        }
-
-        .level {
-          text-transform: none;
-          cursor: pointer;
-          margin-right: 10px;
-          background: $blue-accent;
-          color: white;
-          box-shadow: none;
-          padding: 5px 10px;
-          border-radius: 2px;
-          border: 3px solid transparent;
-          margin-top: 10px;
-          user-select: none;
-          &:hover,
-          &:focus {
-            text-decoration: none;
-            outline: none;
-          }
-        }
-      }
-    }
-    .disclaimer {
-      background: white;
-      box-shadow: 1px 1px 10px 0 rgba(0, 0, 0, 0.05);
-      padding: 20px;
-      margin-bottom: 40px;
-
-      .title {
-        width: 100%;
-        color: $darker-grey;
-        font-size: $larger;
-        padding-left: 10px;
-      }
-
-      p {
-        padding-left: 10px;
-        padding-top: 8px;
-      }
-    }
-
-    .ratings {
-      background: white;
-      border-radius: 3px;
-      padding: 20px;
-      position: relative;
-
-      button {
-        position: absolute;
-        right: 20px;
-        top: 20px;
-      }
-
-      ::ng-deep .title {
-        font-size: $larger;
-      }
-    }
-  }
-
-  .sidebar {
-    width: $sidebar-width;
-    padding: 20px;
-    flex-shrink: 0;
-    margin-top: -190px;
-    z-index: 2;
-    box-sizing: border-box;
-  }
-}
-
-@media (max-width: 800px) {
-  .flex {
-    flex-direction: column-reverse !important;
-
-    .sidebar {
-      margin: auto;
-      margin-top: -50px !important;
-      width: 100% !important;
-
-      .btn-group {
-        width: 100% !important;
-        flex-direction: row-reverse;
-        justify-content: space-around;
-
-        .button {
-          width: 48%;
-        }
-      }
-
-      .details-share {
-        max-width: $sidebar-width;
-        margin: auto;
-      }
-    }
-  }
-}
-
-.new-rating-wrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 999999;
-  background: $dark-overlay;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  visibility: hidden;
-  pointer-events: none;
-  opacity: 0;
-  transition: all 0.2s ease;
-
-  &.active {
-    visibility: visible;
-    opacity: 1;
-    pointer-events: all;
-
-    .new-rating {
-      transition-delay: 0.1s;
-      transform: scale(1);
-    }
-  }
-
-  .new-rating {
-    transition-delay: 0s;
-    transition: inherit;
-    transform: scale(0.9);
-    background: white;
-    padding: 15px;
-    min-width: 350px;
-    max-width: 600px;
-    width: 100%;
-    height: 400px;
-    border-radius: 2px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    position: relative;
-
-    .close-new-rating {
-      position: absolute;
-      top: 10px;
-      right: 20px;
-      font-size: $larger;
-      cursor: pointer;
-      color: $medium-grey;
-    }
-
-    .title {
-      font-size: $larger;
-      text-align: center;
-      color: $darker-grey;
-      margin-bottom: 15px;
-    }
-  }
-}
-.inner {
-  background: white;
-  border-radius: 2px;
-  box-shadow: 0 0 12px 1px rgba(0, 0, 0, 0.03);
-  padding: 20px;
-  margin-bottom: 15px;
-
-  .no-revisions {
-    font-size: $normal;
-    color: $dark-grey;
-    padding-right: 10px;
-    text-align: center;
-    display: inline-block;
-  }
-}
-.text {
-  font-size: $normal;
-  color: $dark-grey;
-  padding-right: 10px;
-  padding-left: 50px;
-  display: inline-block;
-}
-@media (max-width: 450px) {
-  .sidebar .btn-group {
-    flex-direction: column !important;
-
-    .button {
-      width: 100% !important;
-    }
-  }
-}
-
-@keyframes newSaveOrDownload {
-  0% {
-    bottom: 0;
-    opacity: 1;
-  }
-  50% {
-    bottom: 10px;
-    opacity: 0;
-  }
-  51% {
-    bottom: -10px;
-    opacity: 0;
-    color: $success-green;
-  }
-  100% {
-    bottom: 0;
-    opacity: 1;
-    color: inherit;
-  }
-}
-
-@keyframes jiggle {
-  0% {
-    transform: scale(1);
-  }
-  25% {
-    transform: scale(0.95);
-  }
-  50% {
-    transform: scale(1.1);
-  }
-  75% {
-    transform: scale(0.95);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-.loading {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  max-width: 300px;
-  font-size: $large;
-  color: $dark-grey;
-  margin: auto;
-  margin-top: 100px;
-
-  .svg-inline--fa {
-    color: $light-blue;
-    font-size: $larger;
-    margin-bottom: 10px;
-  }
-}
