--- src/app/cube/old-details/details.copy.ts
@@ -1,12 +1,0 @@
-export const COPY = {
-    AUTHORPREFIX: 'Created by',
-    CONTRIBUTORSPREFIX: 'Contributors:',
-    DESCRIPTIONTITLE: 'Description',
-    REVIEWMESSAGE: 'This learning object is pending review by the NCCP Review Team and may be downloaded when the review is complete.',
-    PENDINGREVIEW: 'Review Pending',
-    DOWNLOAD: 'Download Now',
-    DOWNLOADING: 'Downloading...',
-    DOWNLOADNOTAVAILABLE: 'Not available for download',
-    SAVE: 'Save to Library',
-    SAVED: 'Saved to your library!'
-};
